Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This effectively breaks reactor tsc stats. Do you have any plans on fixing those?

Yes - it does effectively break the reactor tsc stats.  We're going to need to move the reactor tsc stats into the thread at some point.  That's really where we need the stats, so that we can figure out how much work each thread is doing and know which ones to move to other cores.  I'm not sure it should be this patch though.  So then the question is do we need to fix it before 19.01 releases.

Patch-set: 3
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>