Update patch set 10
Patch Set 10:

I reviewed the spec again:
1  Currently, the spec does not say that for PDU response should be within a time (e.g., 30s)
2 It only describes that when the controller sends termreq, there is a time window to wait for the close the of the fd. If time out reaches, the target closes the socket.

The current issue is that, there is on commands on the admin queue, so causes the timeout. The simplest patch as I submitted is version1, we just bypass the admin qpair.

Patch-set: 10
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>