Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Hi Paul,
> 
> I think that there are two ways:
> 1  doing it in the autogen.sh (but it still has dependency on other libs like nasm, I see your patch)
> 2  Just check the install folder of ISA-L, then we do not need to use submodule. For example, check the ISA-L in /usr/include/isa-l, if the folder exists, we can consider it has ISA-L.
> 
> 2 Just

I think #1 is the way to go (with the check for nasm of course).  The requirements as I understand them wrt what we aer trying to do with ISA-L:

* do not perform and install
* always link to what we've built in the submodule regardless of whether the system has another version of ISA-L on it or not

I think what you have it almost there.  What I did has everything except instead of building it, it installs it which we didn't realize was going to be such a pain until last week.  Jim was going to look at doing the "make and link" patch I *think* - it might be the case that he will have that ready soon and both of us can rebase on top of it for our ISAL needs (CRC for you, DPDK for me)

Thanks!

Patch-set: 3
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>