Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> I reviewed the spec again:
> 1  Currently, the spec does not say that for PDU response should be within a time (e.g., 30s)
> 2 It only describes that when the controller sends termreq, there is a time window to wait for the close the of the fd. If time out reaches, the target closes the socket.
> 
> The current issue is that, there is on commands on the admin queue, so causes the timeout. The simplest patch as I submitted is version1, we just bypass the admin qpair.

I agree that the specification doesn't indicate what to do if the target sends the host a PDU (e.g. R2T) and the host does not respond promptly. However, I thought we had discussed that the best course of action would be to disconnect after 30 seconds so that a malicious or broken host can't continue to consume target resources. It seems to me like the bug is that the timeout while waiting for the PDU response from the host wasn't working correctly and was resulting in disconnects when it shouldn't have (i.e. when a PDU wasn't expected). But your patch here "fixes" that bug by deleting the old timeout and putting in a timeout for an entirely different scenario.

I agree that we need to add a timeout for handling C2HTermReq PDUs to conform to the specification, but that wasn't the bug filed. So go ahead and fix C2HTermReq handling, but separately fix the existing PDU timeout handling so that it correctly only times out connections when it is expecting a PDU response.

Patch-set: 10
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>