Update patch set 10
Patch Set 10:

Hi Ben,

For the timeout of R2T to prevent the malicious connection, for the current code, we still can not prevent. So I would like to revise the commit message but still maintain the current patch. For the R2T case, it is complicated, and would like to solve this late.r

Patch-set: 10
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>