Update patch set 54
Patch Set 54:

> > >
 > > > Thanks for detailed answer.
 > > > Today, size of RDMA SQ is a function of IO depth supported by
 > target. And it's based on assumption that for each rdma request we
 > do maximum single RDMA_READ/RDMA_WRITE operation.
 > > > See max_send_wr attribute in spdk_nvmf_rdma_qpair_initialize
 > function.
 > > > I can't find that this attribute is changed in the patch. I
 > think,instead of factor "2",(SPDK_NVMF_MAX_SGL_ENTRIES + 1) should
 > be used.
 > > > So, from functional point of view, once target support
 > multiple-element SGL, SQ and CQ capacity should be increased. BTW,
 > CQ capacity is a separate issue. Today is hardcoded.
 > > >
 > > > Once, we all queues can handle outstanding IO, we move into
 > scalability area. I think,
 > >
 > > 1. SQ size was a function of I/O depth in the past - This was
 > true, and I had to change the >behavior of the target so that we
 > didn't get errors from ibv_post_send. I opted to solve this >by not
 > growing the SQ, but by keeping track of how many Work requests we
 > had outstanding on >the send queue and holding the requests up in
 > the PENDING_DATA_TRANSFER state until there >were enough to post
 > the entire rdma_req worth of requests to the qpair. Please see
 > >https://review.gerrithub.io/c/spdk/spdk/+/439573 for that
 > implementation. By not increasing >the size of the submission
 > queue, we avoid the scalability problem.
 > 
 > Hi Seth,
 > 
 > At the first glance https://review.gerrithub.io/c/spdk/spdk/+/439573
 > is quite problematic.
 > Let's take an example. Target allows IO depth 5 and initiator
 > actually post 5 IO operations. Each IO operation is multi-element
 > SGL with length 16 (max allowed).
 > SQ length in current approach will be 11 (5*2+1). For handling any
 > IO request, target has to post 16 RDMA requests. How it will work
 > with current approach?
 > Please, correct me if my math is wrong.
 > 
 > 
 > Best regards
 > Sasha

 > > >
 > > > Thanks for detailed answer.
 > > > Today, size of RDMA SQ is a function of IO depth supported by
 > target. And it's based on assumption that for each rdma request we
 > do maximum single RDMA_READ/RDMA_WRITE operation.
 > > > See max_send_wr attribute in spdk_nvmf_rdma_qpair_initialize
 > function.
 > > > I can't find that this attribute is changed in the patch. I
 > think,instead of factor "2",(SPDK_NVMF_MAX_SGL_ENTRIES + 1) should
 > be used.
 > > > So, from functional point of view, once target support
 > multiple-element SGL, SQ and CQ capacity should be increased. BTW,
 > CQ capacity is a separate issue. Today is hardcoded.
 > > >
 > > > Once, we all queues can handle outstanding IO, we move into
 > scalability area. I think,
 > >
 > > 1. SQ size was a function of I/O depth in the past - This was
 > true, and I had to change the >behavior of the target so that we
 > didn't get errors from ibv_post_send. I opted to solve this >by not
 > growing the SQ, but by keeping track of how many Work requests we
 > had outstanding on >the send queue and holding the requests up in
 > the PENDING_DATA_TRANSFER state until there >were enough to post
 > the entire rdma_req worth of requests to the qpair. Please see
 > >https://review.gerrithub.io/c/spdk/spdk/+/439573 for that
 > implementation. By not increasing >the size of the submission
 > queue, we avoid the scalability problem.
 > 
 > Hi Seth,
 > 
 > At the first glance https://review.gerrithub.io/c/spdk/spdk/+/439573
 > is quite problematic.
 > Let's take an example. Target allows IO depth 5 and initiator
 > actually post 5 IO operations. Each IO operation is multi-element
 > SGL with length 16 (max allowed).
 > SQ length in current approach will be 11 (5*2+1). For handling any
 > IO request, target has to post 16 RDMA requests. How it will work
 > with current approach?
 > Please, correct me if my math is wrong.
 > 
 > 
 > Best regards
 > Sasha

Ah, that is a good point. While I am not aware of a target out there that supports such a small I/O depth, it is good to support edge cases. With the current implementation of this patch, we will be able to make forward progress on an I/O queue that supports at least 8 I/O depth. I think in order to support the edge case where we have an I/O depth < 8 I could allocate spdk_max(io_depth * 2 + 1, SPDK_NVMF_MAX_SGL). Then we will always be able to submit at least 1 I/O at a time. Scaling the submission queue up to such a large size will not be necessary since we have the infrastructure in place to throttle if need be.
It might seem like a step backwards to allow (in the worst case) 1 I/O at a time, but we don't actually lose any bandwidth by doing this. Take the example of using our initiator to submit a 32k I/O with 4k SGL elements. In the old implementation, our initiator would split that request into 8 NVMe-oF requests and submit them as 4k I/Os. This would result in 8 RECVs and 8 SENDs at the RDMA layer. In the new case, this would be sent as a single RECV with 8 SGL elements and would result in the same 8 SENDs at the RDMA layer. The benefit is not having to process the splitting code on the initiator and being able to process more I/O on the 7 remaining recv entries that are available to us while we wait for the SEND ops to complete.

Patch-set: 54
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>