Update patch set 54
Patch Set 54:

I do think we need to eventually support initiators with queue depth smaller than the number of NVMe-oF SGL elements we want to support, and we'll need to do this by trickling out a few at a time and maintaining state. However, I think it would be reasonable for this patch to simply check the supported initiator queue depth and then reduce SPDK_NVMF_MAX_SGL down to that value if necessary. The Linux kernel initiator only ever uses 1 element, so I don't think this is a problematic restriction in the short term.

Patch-set: 54