Update patch set 12
Patch Set 12: Code-Review-1

Hi Ziye - I fixed the existing timeout here: https://review.gerrithub.io/c/spdk/spdk/+/441811

If you update this patch to add in the C2HTermReq timeout, without removing the existing timeout or making other changes, I'd happily approve it because I agree that we don't handle C2HTermReq according to the spec.

We can discuss whether we need to be timing out the PDU receive process at all for 19.04. I just did some quick performance tests with and without the timeout, but the results show less than 1% difference, with an error range of about 1%. We'd need much more thorough and controlled performance tests to figure out if it's 0% or actually significant. But even if it doesn't cost any performance, it certainly is extra complication in the code just to catch malfunctioning or malicious initiators. I'm open to the idea that we may not need to do that.

Patch-set: 12
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1