Update patch set 15
Patch Set 15: Code-Review-1

> Patch Set 15:
> 
> > Patch Set 15:
> > 
> > My latest thought is that:
> > 
> > We still make isal as dynamical link library, but if we can copy libisal.so into spdk/build/lib and consolidate the library into build/lib, I think that it should pass the test.
> 
> The thinking here is that whatever we do will be separate from this patch.

Agreed.  I don't want to build ISA-L as a shared library right now.  We've found symbol collisions with Ceph libraries which pulled ISA-L functions as-is, did not change the names of the functions, and did not hide these symbols.  We have ideas using objcopy and symbol prefixing to fix this long term but it means we would not be using the ISA-L built libraries as-is.

Patch-set: 15
Reviewer: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1