Update patch set 5
Patch Set 5: -Code-Review

I've been debating about the direction to take the target regarding SRQ for some time now, and I think I'm now ready to make a decision.

The long term direction of the target is going to be to use the shared receive queue exclusively. It solves several major problems in the current design (CQ resizing, buffer pool allocation, memory scaling).

I'd like to proceed with moving to a shared receive queue in the following steps. For 19.01, we won't merge the SRQ patch due to risk management concerns. But shortly after 19.01 releases I'd like to merge a version of this SRQ patch that does not have the #define - just make it a full conversion to SRQ (I can actually update this patch to do that if you want). Merging SRQ will have a ripple affect across several parts of the code that will then need to get addressed. The major two are 1) the code that dynamically resizes the shared CQ can go away and be based on static math, and 2) the shared buffer pool can go away and we'll just allocate the buffers necessary for each poll group's worst case based on the SRQ size. We'll also need the two patches that follow this one that simplify the shutdown procedure.

I think this is a major design improvement to the target - can't wait to get it all checked in shortly!

Patch-set: 5
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0