Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

Hi Jim,

This patch looks reasonable to me.

This patch is for iSCSI, so I want to share a work I'll do for iSCSI next. LUN ID in R2T response is wrong now and I'll have to fix, but it looks that major initiators don't check it, and so this issue is not critical.

Patch-set: 3
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2