Update patch set 35
Patch Set 35: Code-Review+1

NOTE:  help2man is also a required package for generating the ISAL manual.  Not critical for how we are using it but it does print an error message on make "command not found".  I figured we might want to merge this now and add the new pkgdep after so this gets moving and we also have time to check the pkg name in the TP distros

Patch-set: 35
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1