configure: fix building without isa-l submodule
The isa-l directory is generated automatically by
git on checkout, although it's empty. We check for
that directory existence in order to build ISA-L
and without `git submodule update --init` the build
just fails.

To fix, instead of checking the dir existence, we
check for the autogen.sh file inside that dir.

Change-Id: I7b62513afcd7e9f3953f825c41b265efe3484d57
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>