test/nvmf/nvme-cli: don't run nvme-cli as secondary process
nvme-cli in NVMf tests currently fails to initialize
the NVMe driver and exits straight away [1]. It expects
the primary process to initialize it first, but since the
primary process doesn't operate on any NVMe devices, it
doesn't initialize the driver at all.

Fix this by running nvme-cli without multi-process mode.
It will no longer rely on any external process. In NVMf
tests the nvme-cli is only used to discover, connect and
disconnect from an SPDK NVMf target.

This wasn't an issue before, because we used an outdated
DPDK version for nvme-cli, which didn't detect any other
shared memory processes of SPDK.

[1] nvme.c: 337:nvme_driver_init: *ERROR*: primary process
is not started yet

Change-Id: Id56f94c6655049e87ab9d93ee38853faf40a11e5
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>