Update patch set 7
Patch Set 7: Code-Review-1

Hi guys!
as far as I am aware (just asked on pxt meeting) we're not merging test plans like this, and we'd rather have these descriptions along with the scripts code, so rather in form of comments.
The reason for that is that .md files like this become quickly outdated and at some point everybody starts forgetting to update them (for example we have some test plans in test/vhost which we decided to discard and it STILL needs to be done:))

Patch-set: 7
Reviewer: Gerrit User 1011256 <1011256@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1