Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Hi Ben,
> 
> I prefer not to add this code, but suggesting removing the READY state but directly beginning at NVME_TCP_PDU_RECV_STATE_AWAIT_PDU_CH state. Actually, if we want to add additional code to prevent malicious attack. For example, the initiator only has one bytes send, we can also use the pdu->ch_valid_bytes to check. We know that if ch_valid_bytes==0, there will be no PDU coming even if it is in NVME_TCP_PDU_RECV_STATE_AWAIT_PDU_CH state.

I think it's important to distinguish between a state where the qpair hasn't received any data (and is idle), and a qpair that has received a partial PDU header (and is not idle). This will be important for correctly reporting whether a poll group is "active" in the future.

Patch-set: 2
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>