nvmf/rdma: Defer device initialization late.
This will be used to prevent some potential issues.
if (rqpair->ibv_attr.qp_state == IBV_QPS_ERR ||
rqpair->qpair.state != SPDK_NVMF_QPAIR_ACTIVE)
the rdma request will be set as completed, so we
do not need to access the device too early.

So defer the device initialization, which will be
no harm.

Purpose: To present the unexpected coredump for
real usage.

Change-Id: I6830c9fbc0dba50b19329488c7999dedfc743436
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>