bdev/crypto: fix error path memory leak in driver init
If multiple drivers are being init'd and anything after the first
failed an alloc, we'd fail to free the earlier allocations in the
cleanup path.

The first patch just fixes the bug.  The second refactors driver
init to match review feedback on the compression driver and simplifies
the code for both readability and maintainability.  Plus it matches
how the pending compression driver does init (as much as makes sense).

github issue #633

Change-Id: Ia2d55d9e98fb9470ff8f9b34aeb4ee9f3d0478f5
Signed-off-by: paul luse <paul.e.luse@intel.com>