event/app: add affinity groups
INTRODUCTION
------------

There are cases where we would like to "seclude" a core from general
usage in an SPDK application.

The primary example driving this currently is software based crypto or
compression. Doing crypto or compression in software is fairly
expensive (especially so for compression), so we may want to dedicate
some cores in the application to be devoted only to performing
crypto/compression operations.

This would ensure that iSCSI/NVMf/vhost clients not using encryption
or compression are not negatively affected by these expensive
operations.

IMPLEMENTATION
--------------

Introduce affinity groups allowing to define groups of cores identified
by name.

E.g. use 0-3 lcores for iscsi group and 4-7 to vhost group:

	--cpu-affinity-groups="iscsi@0x0F;vhost@[4-7]"

and get it when needed by:

	struct spdk_cpuset *
	spdk_app_get_affinity_group(const char *name);

In addition function:

	struct spdk_cpuset *
	spdk_app_get_free_core_mask(void);

returns all cores not assigned to any group.

If group is not defined all free cores will be used for this group.

Affinity groups creation/verification rules:

 1) Each defined affinity group is exclusive,
 2) Each not defined group is non-exclusive,
 3) We should grant that there's at least one non-exclusive core,
 4) Master core cannot be assigned to the affinity group

Signed-off-by: Tomasz Kulasek <tomaszx.kulasek@intel.com>
Change-Id: I6a6c5b703d39f453fc971c1201c36dacb0eef61a