nvmf: don't set qpair->group to NULL.
In rdma.c:_nvmf_rdma_disconnect_retry, we have a comment that says
qpair->group is set and never changes after that. Setting the qpair to
NULL hear breaks the assumption of that comment, and can cause a race
condition.
We get a disconnect event on one thread, then poll the completion queue
on a second thread.
While the disconnect event is being processed, we get a recv cpl with an
error code. This causes us to call spdk_nvmf_qpair_disconnect which
calls _spdk_nvmf_qpair_destroy setting the qpair to NULL, then the first
thread calls _nvmf_rdma_qpair_disconnect and sees that qpair->group ==
NULL. This causes us to enter an infinite loop.

Change-Id: I65395d0bbb67edfa7bad2ddc70906606c3d83781
Signed-off-by: Seth Howell <seth.howell@intel.com>